refactor: add service for mutable_tensor
