+++ test.py
@@ -15,13 +15,9 @@ async def work(session):
     tensor = await session.create_mutable_tensor(shape=(100,100,100),dtype=np.double,
     chunksize=(10,10,10),name="mytensor")
     await tensor.write(((11,2,3),(14,5,6),(17,8,9)),1)
-    t = await tensor.read(((11,),(14,),(17,)))
+    await tensor.write(((12,2,3),(15,5,6),(16,8,9)),10)
+    t = await tensor.read(((11,12,2),(14,15,5),(17,16,8)))
     print(t)
-    # t = mt.ones((134217728//1024, 134217728//1024), chunk_size=134217728//16//1024)
-    # info = await session.execute(t)
-    # await info
-    # r = await session.fetch(t)
-    # print(len(r))
 
 async def main():
     client = await new_cluster(n_worker=2,
